Fix:(issue_2066) Remove dependency on golang flag
