+++ flag_bool_with_inverse.go
@@ -148,18 +148,18 @@ func (parent *BoolWithInverseFlag) PostParse() error {
 	return nil
 }
 
-func (parent *BoolWithInverseFlag) ValueToApply() Value {
+func (parent *BoolWithInverseFlag) Set(val string) error {
 	if parent.positiveFlag == nil {
 		parent.initialize()
 	}
 
-	/*if err := parent.positiveFlag.Apply(set); err != nil {
+	if err := parent.positiveFlag.Set(val); err != nil {
 		return err
 	}
 
-	if err := parent.negativeFlag.Apply(set); err != nil {
+	if err := parent.negativeFlag.Set(val); err != nil {
 		return err
-	}*/
+	}
 
 	return nil
 }
