+++ flag_ext.go
@@ -6,12 +6,20 @@ type extFlag struct {
 	f *flag.Flag
 }
 
+func (e *extFlag) PreParse() error {
+	return nil
+}
+
 func (e *extFlag) PostParse() error {
 	return nil
 }
 
-func (e *extFlag) ValueToApply() Value {
-	return nil // e.f.Value
+func (e *extFlag) Set(val string) error {
+	return e.f.Value.Set(val)
+}
+
+func (e *extFlag) Get() any {
+	return e.f.Value.(flag.Getter).Get()
 }
 
 func (e *extFlag) Names() []string {
