+++ flag_impl.go
@@ -19,29 +19,6 @@ type boolFlag interface {
 	IsBoolFlag() bool
 }
 
-type fnValue struct {
-	fn     func(string) error
-	isBool bool
-	v      Value
-}
-
-func (f *fnValue) Get() any           { return f.v.Get() }
-func (f *fnValue) Set(s string) error { return f.fn(s) }
-func (f *fnValue) String() string {
-	if f.v == nil {
-		return ""
-	}
-	return f.v.String()
-}
-
-func (f *fnValue) IsBoolFlag() bool { return f.isBool }
-func (f *fnValue) Count() int {
-	if s, ok := f.v.(Countable); ok {
-		return s.Count()
-	}
-	return 0
-}
-
 // ValueCreator is responsible for creating a flag.Value emulation
 // as well as custom formatting
 //
@@ -142,14 +119,14 @@ func (f *FlagBase[T, C, V]) PostParse() error {
 	if !f.hasBeenSet {
 		if val, source, found := f.Sources.LookupWithSource(); found {
 			if val != "" || reflect.TypeOf(f.Value).Kind() == reflect.String {
-				if err := f.value.Set(val); err != nil {
+				if err := f.Set(val); err != nil {
 					return fmt.Errorf(
 						"could not parse %[1]q as %[2]T value from %[3]s for flag %[4]s: %[5]s",
 						val, f.Value, source, f.Name, err,
 					)
 				}
 			} else if val == "" && reflect.TypeOf(f.Value).Kind() == reflect.Bool {
-				_ = f.value.Set("false")
+				_ = f.Set("false")
 			}
 
 			f.hasBeenSet = true
@@ -159,8 +136,27 @@ func (f *FlagBase[T, C, V]) PostParse() error {
 	return nil
 }
 
-// Apply populates the flag given the flag set and environment
-func (f *FlagBase[T, C, V]) ValueToApply() Value {
+func (f *FlagBase[T, C, V]) PreParse() error {
+	newVal := f.Value
+
+	if f.Destination == nil {
+		f.value = f.creator.Create(newVal, new(T), f.Config)
+	} else {
+		f.value = f.creator.Create(newVal, f.Destination, f.Config)
+	}
+
+	// Validate the given default or values set from external sources as well
+	if f.Validator != nil && f.ValidateDefaults {
+		if err := f.Validator(f.value.Get().(T)); err != nil {
+			return err
+		}
+	}
+	f.applied = true
+	return nil
+}
+
+// Set applies given value from string
+func (f *FlagBase[T, C, V]) Set(val string) error {
 	tracef("apply (flag=%[1]q)", f.Name)
 
 	// TODO move this phase into a separate flag initialization function
@@ -170,50 +166,34 @@ func (f *FlagBase[T, C, V]) ValueToApply() Value {
 	// flag can be applied to different flag sets multiple times while still
 	// keeping the env set.
 	if !f.applied || f.Local {
-		newVal := f.Value
-
-		if f.Destination == nil {
-			f.value = f.creator.Create(newVal, new(T), f.Config)
-		} else {
-			f.value = f.creator.Create(newVal, f.Destination, f.Config)
+		if err := f.PreParse(); err != nil {
+			return err
 		}
+		f.applied = true
+	}
 
-		// Validate the given default or values set from external sources as well
-		if f.Validator != nil && f.ValidateDefaults {
-			if err := f.Validator(f.value.Get().(T)); err != nil {
-				return f.value
-			}
+	if f.count == 1 && f.OnlyOnce {
+		return fmt.Errorf("cant duplicate this flag")
+	}
+
+	f.count++
+	if err := f.value.Set(val); err != nil {
+		return err
+	}
+	f.hasBeenSet = true
+	if f.Validator != nil {
+		if err := f.Validator(f.value.Get().(T)); err != nil {
+			return err
 		}
 	}
+	return nil
+}
 
-	isBool := false
-	if b, ok := f.value.(boolFlag); ok && b.IsBoolFlag() {
-		isBool = true
+func (f *FlagBase[T, C, V]) Get() any {
+	if f.value == nil {
+		f.PreParse()
 	}
-
-	return &fnValue{
-		fn: func(val string) error {
-			if f.count == 1 && f.OnlyOnce {
-				return fmt.Errorf("cant duplicate this flag")
-			}
-			f.count++
-			if err := f.value.Set(val); err != nil {
-				return err
-			}
-			f.hasBeenSet = true
-			if f.Validator != nil {
-				if err := f.Validator(f.value.Get().(T)); err != nil {
-					return err
-				}
-			}
-			return nil
-		},
-		isBool: isBool,
-		v:      f.value,
-	}
-
-	//f.applied = true
-	//return nil
+	return f.value.Get()
 }
 
 // IsDefaultVisible returns true if the flag is not hidden, otherwise false
@@ -283,7 +263,7 @@ func (f *FlagBase[T, C, V]) GetDefaultText() string {
 // RunAction executes flag action if set
 func (f *FlagBase[T, C, V]) RunAction(ctx context.Context, cmd *Command) error {
 	if f.Action != nil {
-		return f.Action(ctx, cmd, cmd.Value(f.Name).(T))
+		return f.Action(ctx, cmd, f.value.Get().(T))
 	}
 
 	return nil
@@ -303,4 +283,18 @@ func (f *FlagBase[T, C, VC]) IsMultiValueFlag() bool {
 // IsLocal returns false if flag needs to be persistent across subcommands
 func (f *FlagBase[T, C, VC]) IsLocal() bool {
 	return f.Local
+}
+
+// IsBoolFlag returns whether the flag doesnt need to accept args
+func (f *FlagBase[T, C, VC]) IsBoolFlag() bool {
+	if f.value == nil {
+		f.PreParse()
+	}
+	bf, ok := f.value.(boolFlag)
+	return ok && bf.IsBoolFlag()
+}
+
+// Count returns the number of times this flag has been invoked
+func (f *FlagBase[T, C, VC]) Count() int {
+	return f.count
 }
