+++ help.go
@@ -214,7 +214,7 @@ func printFlagSuggestions(lastArg string, flags []Flag, writer io.Writer) {
 			continue
 		}
 		// match if last argument matches this flag and it is not repeated
-		if strings.HasPrefix(name, cur) && cur != name && !cliArgContains(name, os.Args) {
+		if strings.HasPrefix(name, cur) && cur != name /* && !cliArgContains(name, os.Args)*/ {
 			flagCompletion := fmt.Sprintf("%s%s", strings.Repeat("-", count), name)
 			if usage != "" && strings.HasSuffix(os.Getenv("SHELL"), "zsh") {
 				flagCompletion = fmt.Sprintf("%s:%s", flagCompletion, usage)
@@ -225,7 +225,10 @@ func printFlagSuggestions(lastArg string, flags []Flag, writer io.Writer) {
 }
 
 func DefaultCompleteWithFlags(ctx context.Context, cmd *Command) {
-	args := os.Args
+	args := cmd.inputArgs
+	if args == nil {
+		args = []string{}
+	}
 	if cmd != nil && cmd.parent != nil {
 		args = cmd.Args().Slice()
 		tracef("running default complete with flags[%v] on command %[2]q", args, cmd.Name)
